Move steve specific (resourceField) code to steve models - create models for steve schemas and apply to cluster and management stores - move resoureField based validationto steve model - move pathExistsInSchema to steve store getter - don't fetch schemaDefinitions on start up when saving prefs (not needed and blocking)
