+++ shell/components/form/ResourceTabs/index.vue
@@ -113,7 +113,7 @@ export default {
         return false;
       }
 
-      return this.$store.getters[`${ this.inStore }/pathExistsInSchema`](this.value.type, 'status.conditions');
+      return this.$store.getters[`${ this.inStore }/pathExistsInSchema`](this.value.type, 'status.conditions'); // TODO: RC test in norman resources (moved pathExistsInSchema to steve getters)???
     },
 
     showEvents() {
