+++ shell/models/cluster/schema.js
@@ -1,0 +1,6 @@
+import SteveSchema from '../steve-schema';
+
+/**
+ * This is the steve schema that's used for the cluster store
+ */
+export default class ClusterSchema extends SteveSchema { }
