+++ shell/models/management/schema.js
@@ -1,0 +1,6 @@
+import SteveSchema from '../steve-schema';
+
+/**
+ * This is the steve schema that's used for the management store
+ */
+export default class ManagementSchema extends SteveSchema { }
