+++ shell/plugins/steve/getters.js
@@ -217,4 +217,34 @@ export default {
     return cache.generation;
   },
 
+  pathExistsInSchema: (state, getters) => (type, path) => {
+    let schema = getters.schemaFor(type);
+    const schemaDefinitions = schema.schemaDefinitions;
+    const parts = splitObjectPath(path);
+
+    while ( parts.length ) {
+      const key = parts.shift();
+
+      const field = schema.resourceFields?.[key];
+
+      type = field?.type;
+
+      if ( !type ) {
+        return false;
+      }
+
+      if ( parts.length ) {
+        type = parseType(type, field).pop(); // Get the main part of array[map[something]] => something
+
+        schema = schemaDefinitions.others[type];
+
+        if ( !schema ) {
+          return false;
+        }
+      }
+    }
+
+    return true;
+  },
+
 };
